Add WebsocketServer to main execution
